refactor: introduce Writer interface and extract client/geth/
Foundation refactor for the multi-client architecture (PR#1 of 3 in the
deep-feature-planning plan at ~/.claude/plans/do-you-recall-the-gleaming-fox.md).

* generator.Writer interface + WriterFactory + RegisterDefaultWriterFactory
  abstract the geth-specific writer surface that Generator drives. New() now
  resolves the registered default factory; NewWithWriter accepts an explicit
  one for tests / future clients.
* GethWriter and its key encoders move to client/geth/ as geth.Writer +
  package-internal helpers. genesis.WriteGenesisBlock + the SnapshotGenerator
  + prefixWriter machinery move alongside as geth.WriteGenesisBlock — all
  geth-specific PathDB metadata work now lives under client/geth/.
* genesis package retains client-neutral parsers only (LoadGenesis,
  ToStateAccounts, GetAllocStorage, GetAllocCode).
* client/geth.init() registers itself as the default factory; main.go and
  e2e_test.go opt in via named import. Internal generator tests register
  via setup_external_test.go (package generator_test) to avoid the
  generator -> client/geth -> generator import cycle.
* docs/ARCHITECTURE.md updated with a Client Adapters section and the new
  client/geth/ tree layout.

Behavior-preserving:
* Same --seed produces byte-identical state roots before vs. after (smoke:
  both runs of accounts=50/contracts=10/seed=42 -> 0xfcbc88712c97e3d3...).
* Reproducibility tests + bintrie golden hash 0xee656cf3... still pass.
* go test -race ./generator/... ./genesis/... ./client/... all pass.
