+++ client/geth/key_encoding_test.go
@@ -1,0 +1,46 @@
+package geth
+
+import (
+	"testing"
+
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+)
+
+// TestKeyEncoding pins the byte layout of the snapshot-layer keys
+// (a-prefix accounts, o-prefix storage, c-prefix code) that the geth
+// writer emits. Geth's rawdb readers expect this exact format.
+func TestKeyEncoding(t *testing.T) {
+	addr := common.HexToAddress("0x1234567890123456789012345678901234567890")
+	addrHash := crypto.Keccak256Hash(addr[:])
+
+	// Test account snapshot key
+	accKey := accountSnapshotKey(addrHash)
+	if len(accKey) != 1+common.HashLength {
+		t.Errorf("Account key wrong length: got %d, want %d", len(accKey), 1+common.HashLength)
+	}
+	if accKey[0] != 'a' {
+		t.Errorf("Account key wrong prefix: got %c, want 'a'", accKey[0])
+	}
+
+	// Test storage snapshot key
+	storageKey := common.HexToHash("0xabcdef")
+	storageKeyHash := crypto.Keccak256Hash(storageKey[:])
+	stoKey := storageSnapshotKey(addrHash, storageKeyHash)
+	if len(stoKey) != 1+common.HashLength*2 {
+		t.Errorf("Storage key wrong length: got %d, want %d", len(stoKey), 1+common.HashLength*2)
+	}
+	if stoKey[0] != 'o' {
+		t.Errorf("Storage key wrong prefix: got %c, want 'o'", stoKey[0])
+	}
+
+	// Test code key
+	codeHash := crypto.Keccak256Hash([]byte("test code"))
+	cKey := codeKey(codeHash)
+	if len(cKey) != 1+common.HashLength {
+		t.Errorf("Code key wrong length: got %d, want %d", len(cKey), 1+common.HashLength)
+	}
+	if cKey[0] != 'c' {
+		t.Errorf("Code key wrong prefix: got %c, want 'c'", cKey[0])
+	}
+}
