+++ client/geth/writer.go
@@ -1,4 +1,4 @@
-package generator
+package geth
 
 import (
 	"fmt"
@@ -13,101 +13,96 @@ import (
 	"github.com/ethereum/go-ethereum/ethdb"
 	"github.com/ethereum/go-ethereum/ethdb/pebble"
 	"github.com/ethereum/go-ethereum/rlp"
+
+	"github.com/nerolation/state-actor/generator"
 	"github.com/nerolation/state-actor/genesis"
 )
 
-// WriterStats holds cumulative byte counts from state writes.
-type WriterStats struct {
-	AccountBytes uint64
-	StorageBytes uint64
-	CodeBytes    uint64
-}
-
-// GethWriter writes state to a geth-compatible Pebble database.
+// Writer writes state to a geth-compatible Pebble database.
 // It uses the snapshot layer format with hashed keys.
-type GethWriter struct {
+type Writer struct {
 	db        ethdb.KeyValueStore
 	dbPath    string
 	batchSize int
 	workers   int
 
-	bw *gethBatchWriter
+	bw *batchWriter
 
 	accountBytes atomic.Uint64
 	storageBytes atomic.Uint64
 	codeBytes    atomic.Uint64
 }
 
-// NewGethWriter creates a new geth-format state writer.
-func NewGethWriter(dbPath string, batchSize, workers int) (*GethWriter, error) {
+// NewWriter creates a new geth-format state writer.
+func NewWriter(dbPath string, batchSize, workers int) (*Writer, error) {
 	db, err := pebble.New(dbPath, 512, 256, "stategen/", false)
 	if err != nil {
 		return nil, fmt.Errorf("failed to open pebble database: %w", err)
 	}
 
-	w := &GethWriter{
+	w := &Writer{
 		db:        db,
 		dbPath:    dbPath,
 		batchSize: batchSize,
 		workers:   workers,
 	}
 
-	w.bw = newGethBatchWriter(db, batchSize, workers)
+	w.bw = newBatchWriter(db, batchSize, workers)
 
 	return w, nil
 }
 
 // DB returns the underlying database for external use (e.g., genesis writing).
-func (w *GethWriter) DB() ethdb.KeyValueStore {
+func (w *Writer) DB() ethdb.KeyValueStore {
 	return w.db
 }
 
 // WriteAccount writes an account to the snapshot layer.
 // addrHash is pre-computed keccak256(addr) to avoid redundant hashing.
-func (w *GethWriter) WriteAccount(addr common.Address, addrHash common.Hash, acc *types.StateAccount, incarnation uint64) error {
+func (w *Writer) WriteAccount(addr common.Address, addrHash common.Hash, acc *types.StateAccount, incarnation uint64) error {
 	slimData := types.SlimAccountRLP(*acc)
-	key := gethAccountSnapshotKey(addrHash)
+	key := accountSnapshotKey(addrHash)
 	return w.bw.put(key, slimData, &w.accountBytes)
 }
 
 // WriteStorage writes a storage slot to the snapshot layer.
 // addrHash and slotHash are pre-computed keccak256 hashes (addr and slot are unused in geth format).
-func (w *GethWriter) WriteStorage(addr common.Address, addrHash common.Hash, slot common.Hash, slotHash common.Hash, value common.Hash) error {
-	valueRLP, err := gethEncodeStorageValue(value)
+func (w *Writer) WriteStorage(addr common.Address, addrHash common.Hash, slot common.Hash, slotHash common.Hash, value common.Hash) error {
+	valueRLP, err := encodeStorageValue(value)
 	if err != nil {
 		return fmt.Errorf("encode storage value: %w", err)
 	}
-	key := gethStorageSnapshotKey(addrHash, slotHash)
+	key := storageSnapshotKey(addrHash, slotHash)
 	return w.bw.put(key, valueRLP, &w.storageBytes)
 }
 
 // WriteStorageRLP writes a storage slot with pre-encoded RLP value.
 // Avoids double-encoding when the caller already has the RLP bytes.
-func (w *GethWriter) WriteStorageRLP(addrHash common.Hash, slotHash common.Hash, valueRLP []byte) error {
-	key := gethStorageSnapshotKey(addrHash, slotHash)
+func (w *Writer) WriteStorageRLP(addrHash common.Hash, slotHash common.Hash, valueRLP []byte) error {
+	key := storageSnapshotKey(addrHash, slotHash)
 	return w.bw.put(key, valueRLP, &w.storageBytes)
 }
 
 // WriteRawStorage writes a storage slot using a pre-hashed trie key.
 // The hashedSlot bypasses keccak256 and is used directly as the snapshot key.
-func (w *GethWriter) WriteRawStorage(addr common.Address, incarnation uint64, hashedSlot, value common.Hash) error {
+func (w *Writer) WriteRawStorage(addr common.Address, incarnation uint64, hashedSlot, value common.Hash) error {
 	addrHash := crypto.Keccak256Hash(addr[:])
-	valueRLP, err := gethEncodeStorageValue(value)
+	valueRLP, err := encodeStorageValue(value)
 	if err != nil {
 		return fmt.Errorf("encode storage value: %w", err)
 	}
-	key := gethStorageSnapshotKey(addrHash, hashedSlot)
+	key := storageSnapshotKey(addrHash, hashedSlot)
 	return w.bw.put(key, valueRLP, &w.storageBytes)
 }
 
 // WriteCode writes contract bytecode.
-func (w *GethWriter) WriteCode(codeHash common.Hash, code []byte) error {
-	key := gethCodeKey(codeHash)
+func (w *Writer) WriteCode(codeHash common.Hash, code []byte) error {
+	key := codeKey(codeHash)
 	return w.bw.put(key, code, &w.codeBytes)
 }
 
 // SetStateRoot writes the snapshot root marker.
-func (w *GethWriter) SetStateRoot(root common.Hash) error {
+func (w *Writer) SetStateRoot(root common.Hash) error {
 	if err := w.db.Put([]byte("SnapshotRoot"), root[:]); err != nil {
 		return err
 	}
@@ -119,9 +114,9 @@ func (w *GethWriter) SetStateRoot(root common.Hash) error {
 	rawdb.WritePersistentStateID(w.db, 0)
 	rawdb.WriteSnapshotRoot(w.db, root)
 	// Mark the snapshot generator as Done so geth doesn't try to regenerate
-	// the snapshot from scratch on first open. See genesis.WriteCompletedSnapshotGenerator
+	// the snapshot from scratch on first open. See WriteCompletedSnapshotGenerator
 	// for the full rationale.
-	if err := genesis.WriteCompletedSnapshotGenerator(w.db); err != nil {
+	if err := WriteCompletedSnapshotGenerator(w.db); err != nil {
 		return fmt.Errorf("write snapshot generator: %w", err)
 	}
 	return nil
@@ -129,7 +124,7 @@ func (w *GethWriter) SetStateRoot(root common.Hash) error {
 
 // Flush commits all pending writes and closes the async batch pipeline.
 // This is a shutdown-once operation — don't call it mid-run.
-func (w *GethWriter) Flush() error {
+func (w *Writer) Flush() error {
 	return w.bw.finish()
 }
 
@@ -139,19 +134,19 @@ func (w *GethWriter) Flush() error {
 // mid-generation (e.g. to force a dirSize sample to see the latest bytes).
 // The caller is responsible for coordinating that all desired Write* calls
 // have already returned before flushing.
-func (w *GethWriter) FlushBatch() error {
+func (w *Writer) FlushBatch() error {
 	return w.bw.flushAndDrainSync()
 }
 
 // Close closes the writer.
-func (w *GethWriter) Close() error {
+func (w *Writer) Close() error {
 	w.bw.close()
 	return w.db.Close()
 }
 
 // Stats returns write statistics.
-func (w *GethWriter) Stats() WriterStats {
-	return WriterStats{
+func (w *Writer) Stats() generator.WriterStats {
+	return generator.WriterStats{
 		AccountBytes: w.accountBytes.Load(),
 		StorageBytes: w.storageBytes.Load(),
 		CodeBytes:    w.codeBytes.Load(),
@@ -159,40 +154,40 @@ func (w *GethWriter) Stats() WriterStats {
 }
 
 // WriteGenesisBlockFull writes the genesis block with full genesis config.
-func (w *GethWriter) WriteGenesisBlockFull(genesisConfig *genesis.Genesis, stateRoot common.Hash, binaryTrie bool) error {
+func (w *Writer) WriteGenesisBlockFull(genesisConfig *genesis.Genesis, stateRoot common.Hash, binaryTrie bool) error {
 	ancientDir := filepath.Join(w.dbPath, "ancient")
-	_, err := genesis.WriteGenesisBlock(w.db, genesisConfig, stateRoot, binaryTrie, ancientDir)
+	_, err := WriteGenesisBlock(w.db, genesisConfig, stateRoot, binaryTrie, ancientDir)
 	return err
 }
 
 // --- Batch writer for parallel writes ---
 
-type gethBatchWriter struct {
+type batchWriter struct {
 	db        ethdb.KeyValueStore
 	batchSize int
-	batchChan chan *gethBatchWork
+	batchChan chan *batchWork
 	errChan   chan error
 	wg        sync.WaitGroup
 	closeOnce sync.Once
 	// mu serialises put() (hot path, single-goroutine in normal operation)
 	// with mid-run flush() calls issued from a different goroutine (e.g.
 	// target-size size sampling). The async batch-commit workers don't
-	// touch bw.batch directly — they consume the detached *gethBatchWork
+	// touch bw.batch directly — they consume the detached *batchWork
 	// sent on batchChan — so they don't need to hold this mutex.
 	mu    sync.Mutex
 	batch ethdb.Batch
 	count int
 }
 
-type gethBatchWork struct {
+type batchWork struct {
 	batch ethdb.Batch
 }
 
-func newGethBatchWriter(db ethdb.KeyValueStore, batchSize, workers int) *gethBatchWriter {
-	bw := &gethBatchWriter{
+func newBatchWriter(db ethdb.KeyValueStore, batchSize, workers int) *batchWriter {
+	bw := &batchWriter{
 		db:        db,
 		batchSize: batchSize,
-		batchChan: make(chan *gethBatchWork, workers*2),
+		batchChan: make(chan *batchWork, workers*2),
 		errChan:   make(chan error, 1),
 		batch:     db.NewBatch(),
 	}
@@ -216,7 +211,7 @@ func newGethBatchWriter(db ethdb.KeyValueStore, batchSize, workers int) *gethBat
 	return bw
 }
 
-func (bw *gethBatchWriter) put(key, value []byte, counter *atomic.Uint64) error {
+func (bw *batchWriter) put(key, value []byte, counter *atomic.Uint64) error {
 	bw.mu.Lock()
 	if err := bw.batch.Put(key, value); err != nil {
 		bw.mu.Unlock()
@@ -234,18 +229,18 @@ func (bw *gethBatchWriter) put(key, value []byte, counter *atomic.Uint64) error 
 
 // flushExternal is the public-facing flush entry. flushLocked expects the
 // caller to already hold bw.mu.
-func (bw *gethBatchWriter) flushExternal() error {
+func (bw *batchWriter) flushExternal() error {
 	bw.mu.Lock()
 	defer bw.mu.Unlock()
 	return bw.flushLocked()
 }
 
-func (bw *gethBatchWriter) flushLocked() error {
+func (bw *batchWriter) flushLocked() error {
 	if bw.count == 0 {
 		return nil
 	}
 	select {
-	case bw.batchChan <- &gethBatchWork{batch: bw.batch}:
+	case bw.batchChan <- &batchWork{batch: bw.batch}:
 	case err := <-bw.errChan:
 		return fmt.Errorf("batch worker failed: %w", err)
 	}
@@ -256,7 +251,7 @@ func (bw *gethBatchWriter) flushLocked() error {
 
 // flush is retained as the lock-acquiring form used by external callers
 // (FlushBatch and finish).
-func (bw *gethBatchWriter) flush() error {
+func (bw *batchWriter) flush() error {
 	return bw.flushExternal()
 }
 
@@ -265,7 +260,7 @@ func (bw *gethBatchWriter) flush() error {
 // the call returns. It swaps in a fresh batch under the lock, then
 // commits the old one directly; the async workers continue handling
 // their own queued batches normally.
-func (bw *gethBatchWriter) flushAndDrainSync() error {
+func (bw *batchWriter) flushAndDrainSync() error {
 	bw.mu.Lock()
 	if bw.count == 0 {
 		bw.mu.Unlock()
@@ -278,7 +273,7 @@ func (bw *gethBatchWriter) flushAndDrainSync() error {
 	return oldBatch.Write()
 }
 
-func (bw *gethBatchWriter) finish() error {
+func (bw *batchWriter) finish() error {
 	if err := bw.flush(); err != nil {
 		return err
 	}
@@ -293,7 +288,7 @@ func (bw *gethBatchWriter) finish() error {
 	return nil
 }
 
-func (bw *gethBatchWriter) close() {
+func (bw *batchWriter) close() {
 	bw.closeOnce.Do(func() { close(bw.batchChan) })
 	bw.wg.Wait()
 }
@@ -301,29 +296,29 @@ func (bw *gethBatchWriter) close() {
 // --- Key encoding functions matching geth's rawdb schema ---
 
 var (
-	gethSnapshotAccountPrefix = []byte("a")
-	gethSnapshotStoragePrefix = []byte("o")
-	gethCodePrefix            = []byte("c")
+	snapshotAccountPrefix = []byte("a")
+	snapshotStoragePrefix = []byte("o")
+	codePrefix            = []byte("c")
 )
 
-func gethAccountSnapshotKey(hash common.Hash) []byte {
-	return append(gethSnapshotAccountPrefix, hash.Bytes()...)
+func accountSnapshotKey(hash common.Hash) []byte {
+	return append(snapshotAccountPrefix, hash.Bytes()...)
 }
 
-func gethStorageSnapshotKey(accountHash, storageHash common.Hash) []byte {
-	buf := make([]byte, len(gethSnapshotStoragePrefix)+common.HashLength+common.HashLength)
-	n := copy(buf, gethSnapshotStoragePrefix)
+func storageSnapshotKey(accountHash, storageHash common.Hash) []byte {
+	buf := make([]byte, len(snapshotStoragePrefix)+common.HashLength+common.HashLength)
+	n := copy(buf, snapshotStoragePrefix)
 	n += copy(buf[n:], accountHash.Bytes())
 	copy(buf[n:], storageHash.Bytes())
 	return buf
 }
 
-func gethCodeKey(hash common.Hash) []byte {
-	return append(gethCodePrefix, hash.Bytes()...)
+func codeKey(hash common.Hash) []byte {
+	return append(codePrefix, hash.Bytes()...)
 }
 
-func gethEncodeStorageValue(value common.Hash) ([]byte, error) {
-	trimmed := gethTrimLeftZeroes(value[:])
+func encodeStorageValue(value common.Hash) ([]byte, error) {
+	trimmed := trimLeftZeroes(value[:])
 	if len(trimmed) == 0 {
 		return nil, nil
 	}
@@ -334,7 +329,7 @@ func gethEncodeStorageValue(value common.Hash) ([]byte, error) {
 	return encoded, nil
 }
 
-func gethTrimLeftZeroes(s []byte) []byte {
+func trimLeftZeroes(s []byte) []byte {
 	for i, v := range s {
 		if v != 0 {
 			return s[i:]
