+++ generator/generator_test.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/rawdb"
-	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/ethdb/pebble"
 	"github.com/ethereum/go-ethereum/rlp"
 )
@@ -252,41 +251,6 @@ func TestStorageValueEncoding(t *testing.T) {
 				t.Errorf("Encoding mismatch for %s: got %x, want %x", tc.name, encoded, tc.expected)
 			}
 		})
-	}
-}
-
-func TestKeyEncoding(t *testing.T) {
-	addr := common.HexToAddress("0x1234567890123456789012345678901234567890")
-	addrHash := crypto.Keccak256Hash(addr[:])
-
-	// Test account snapshot key
-	accKey := gethAccountSnapshotKey(addrHash)
-	if len(accKey) != 1+common.HashLength {
-		t.Errorf("Account key wrong length: got %d, want %d", len(accKey), 1+common.HashLength)
-	}
-	if accKey[0] != 'a' {
-		t.Errorf("Account key wrong prefix: got %c, want 'a'", accKey[0])
-	}
-
-	// Test storage snapshot key
-	storageKey := common.HexToHash("0xabcdef")
-	storageKeyHash := crypto.Keccak256Hash(storageKey[:])
-	stoKey := gethStorageSnapshotKey(addrHash, storageKeyHash)
-	if len(stoKey) != 1+common.HashLength*2 {
-		t.Errorf("Storage key wrong length: got %d, want %d", len(stoKey), 1+common.HashLength*2)
-	}
-	if stoKey[0] != 'o' {
-		t.Errorf("Storage key wrong prefix: got %c, want 'o'", stoKey[0])
-	}
-
-	// Test code key
-	codeHash := crypto.Keccak256Hash([]byte("test code"))
-	cKey := gethCodeKey(codeHash)
-	if len(cKey) != 1+common.HashLength {
-		t.Errorf("Code key wrong length: got %d, want %d", len(cKey), 1+common.HashLength)
-	}
-	if cKey[0] != 'c' {
-		t.Errorf("Code key wrong prefix: got %c, want 'c'", cKey[0])
 	}
 }
